feat: Move internal libraries to open-source
