+++ packages/react/lottie/src/LottieRef.ts
@@ -1,0 +1,6 @@
+export interface LottieRef {
+  start: () => void;
+  stop: () => void;
+  play: () => void;
+  pause: () => void;
+}
