+++ packages/react/use-funnel/src/useFunnel.tsx
@@ -1,18 +1,19 @@
 /** @tossdocs-ignore */
+
 import { assert } from '@toss/assert';
 import { safeSessionStorage } from '@toss/storage';
 import { useQueryParam } from '@toss/use-query-param';
 import { QS } from '@toss/utils';
 import deepEqual from 'fast-deep-equal';
-import { useRouter } from 'next/router';
-import { SetStateAction, useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { useQuery } from 'react-query';
+import { useRouter } from 'next/router.js';
+import { useCallback, useEffect, useMemo, useRef } from 'react';
 import { Funnel, FunnelProps, Step, StepProps } from './Funnel';
 import { NonEmptyArray } from './models';
 
 interface SetStepOptions {
   stepChangeType?: 'push' | 'replace';
   preserveQuery?: boolean;
+  query?: Record<string, any>;
 }
 
 type RouteFunnelProps<Steps extends NonEmptyArray<string>> = Omit<FunnelProps<Steps>, 'steps' | 'step'>;
@@ -38,14 +39,14 @@ export const useFunnel = <Steps extends NonEmptyArray<string>>(
   withState: <StateExcludeStep extends Record<string, unknown> & { step?: never }>(
     initialState: StateExcludeStep
   ) => [
-    FunnelComponent<Steps>,
-    StateExcludeStep,
-    (
-      next:
-        | Partial<StateExcludeStep & { step: Steps[number] }>
-        | ((next: Partial<StateExcludeStep & { step: Steps[number] }>) => StateExcludeStep & { step: Steps[number] })
-    ) => void
-  ];
+      FunnelComponent<Steps>,
+      StateExcludeStep,
+      (
+        next:
+          | Partial<StateExcludeStep & { step: Steps[number] }>
+          | ((next: Partial<StateExcludeStep & { step: Steps[number] }>) => StateExcludeStep & { step: Steps[number] })
+      ) => void
+    ];
 } => {
   const router = useRouter();
   const stepQueryKey = options?.stepQueryKey ?? DEFAULT_STEP_QUERY_KEY;
@@ -76,10 +77,11 @@ export const useFunnel = <Steps extends NonEmptyArray<string>>(
 
   const setStep = useCallback(
     (step: Steps[number], setStepOptions?: SetStepOptions) => {
-      const { preserveQuery = true } = setStepOptions ?? {};
+      const { preserveQuery = true, query = {} } = setStepOptions ?? {};
 
       const url = `${QS.create({
         ...(preserveQuery ? router.query : undefined),
+        ...query,
         [stepQueryKey]: step,
       })}`;
 
@@ -154,21 +156,21 @@ export const useFunnel = <Steps extends NonEmptyArray<string>>(
     return [FunnelComponent, state, setState] as const;
   }
 
-  return Object.assign([FunnelComponent, setStep] as const, { withState }) as readonly [
+  return Object.assign([FunnelComponent, setStep] as const, { withState }) as unknown as readonly [
     FunnelComponent<Steps>,
-    (step: Steps[number], options?: SetStepOptions) => void
+    (step: Steps[number], options?: SetStepOptions) => Promise<void>
   ] & {
     withState: <StateExcludeStep extends Record<string, unknown> & { step?: never }>(
       initialState: StateExcludeStep
     ) => [
-      FunnelComponent<Steps>,
-      StateExcludeStep,
-      (
-        next:
-          | Partial<StateExcludeStep & { step: Steps[number] }>
-          | ((next: Partial<StateExcludeStep & { step: Steps[number] }>) => StateExcludeStep & { step: Steps[number] })
-      ) => void
-    ];
+        FunnelComponent<Steps>,
+        StateExcludeStep,
+        (
+          next:
+            | Partial<StateExcludeStep & { step: Steps[number] }>
+            | ((next: Partial<StateExcludeStep & { step: Steps[number] }>) => StateExcludeStep & { step: Steps[number] })
+        ) => void
+      ];
   };
 };
